requirements-dev.txt
setuptools>=65.0.0
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.0
django-filter==23.5
python-decouple==3.8
Pillow>=10.1.0
drf-spectacular==0.26.5

# Frontend
django-tailwind==3.8.0
django-htmx==1.17.2
django-widget-tweaks==1.5.0
